+++ test_spuce/cfft.cpp
@@ -24,7 +24,7 @@ cfft::cfft(size_t size, float_type scalef1, float_type scalef2, float_type scale
     if (((size_t)1 << k) == size) break;
     if (k == 18 || ((size_t)1 << k) > size) throw "cfft: size not power of 2";
   }
-  N = 1 << k;
+  N = (size_t)1 << k;
   log2N = k;
 
   bitrev = new int[N];
